이아림 6주-2 머지 요청드립니다

+++ src/main/java/com/tn/member/dao/MemberDAOImpl.java
@@ -189,7 +189,7 @@ public class MemberDAOImpl implements MemberDAO  {
 	}
 
 	@Override
-	public double getPointRate(String userId) throws Exception {
+	public int getPointRate(String userId) throws Exception {
 		
 		return sess.selectOne(NS + ".getPointRate", userId);
 	}
